// src/app/compartido/diseno/enums/estilo-del-texto.enum.ts
// Estilos del texto
export enum EstilosDelTexto {
    REGULAR = 'regular',
    BOLD = 'bold',
    ITALIC = 'italic',
    REGULARCONSOMBRA = 'regularSombra',
    BOLDCONSOMBRA = 'boldSombra',
    REGULARENITALIC = 'regularItalic',
    BOLDENITALIC = 'boldItalic',
    REGULARENITALICCONSOMBRA = 'regularItalicSombra',
    BOLDENITALICCONSOMBRA = 'boldItalicSombra',
}