// src/app/compartido/diseno/enums/tamano-item-menu.enum.ts
// El uso define el tamano y demas propiedades que debe llevar
export enum TamanoItemMenu {
    ITEMMENUCREARPERFIL = 'item-menu-alto-3774', // Para el menu de elegir el perfil para crearlo
    ITEMMENUELEGIRPERFILINFO = 'item-menu-alto-1948', // Para el menu de elegir el perfil antes de acceder al menu o home de la app, al inicio suele aparece un item informativo con las acciones y demas. El tamano del mismo es diferente
    ITEMMENUELEGIRPERFIL = 'item-menu-alto-4255', // Para el menu de elegir el perfil antes de acceder al menu o home de la app
    ITEM_MENU_GENERAL = 'item-menu-alto-2621' // Para el item del menu o home de la app    
}